# Fixed external Browserify bundle