The current vendorJS task creates a vendor.js file with all of the libraries, but then Browserify also bundles the libraries into the bundle.js file during the current bundle-dev/bundle-prod tasks. This change alters the gulp tasks so that bundle.js does not contain libraries which were bundled into vendor.js, speeding up the bundle-dev/bundle-prod tasks as I expect was the original intention.