@goodbomb 2015-04-06
Hi, thanks for the contribution, however I don't believe you're using the external vendor files as intended. The vendor.js file is meant specifically for external libraries that can't be used with Angular's dependency injection and thus have to be called through the index.html file (such as Bootstrap). I may not have explained it clearly in the README, but the vendor.js file isn't meant for modules that can use angular dependency injection (it's also possible that the libraries I've used in this demo no longer require this approach).

Currently, the vendor files don't get added twice to the Bundle.js file (I've checked). You may be coming at this from the grunt.js approach of having to explicitly define each vendor file that will be included in the browserify bundle, but you don't have to do that in this case. The way that this gulpfile works is that it bundles up all of the dependencies injected throughout the application and ignores any vendor files in the libs folder that haven't been injected into the app. That way you don't have to manually remove and add each dependency in your gulpfile.

That being said, I'm open to exploring your solution further. My immediate concern has to do with angular's dependency injection. Does your change mean that when you require a dependency in the angular app, it won't be included by browserify if it's already in the vendor.js file?