@goodbomb 2015-04-06
Also here is the note at the end of the README under the section I mentioned:

> NOTE: This is meant strictly for third party libraries that cannot be installed using npm install or bower install. You should use one of those two methods primarily for installing third party libraries so that you can easily inject them into your modules.