@leftiness 2015-04-06
I suppose I should have known better than to assume that my "problem" was also a "problem" to you. From your perspective, this is a proposal to change the functionality of the `vendor.js` file, so I should have written a bit more than I originally did.

#### Let's start with a few numbers. 

These numbers are for my PC, an old gaming tower with an AMD Phenom II X6 at stock frequency.

- Your branch
	- Original `gulp build`: ~18s
	- Automatic rebundle after making a change to a JavaScript file: ~10s
	- `gulp build-dev`: ~13s
- My branch
	- Original `gulp build`: ~33s (~14s <sup>[1]</sup>)
	- Automatic rebundle after making a change to a JavaScript file: ~5s
	- `gulp build-dev`: ~8s

#### Now an explanation. 

I read about an interesting Browserify concept: external bundles <sup>[2][3][4][5]</sup>. I think Ben Clinkinbeard <sup>[3]</sup> summarized the concept well:

> Among the handful of options Browserify supports, one of them is external. This is simply a way to tell Browserify "Hey, when you see require() statement looking for X, don't worry about finding it in the tree. It will be available, trust me."

Instead of bundling up all of the libraries with the application code, the change that I proposed with this pull request creates two separate bundles. One bundle (`vendor.js`) contains libraries like AngularJS. The other (`bundle.js`) contains application code. In this way, the `build-dev` task has less to do when you change something in your application code. You have to add each dependency to the gulpfile, but every build thereafter executes significantly faster. 

#### In conclusion...

I debated whether I should email first or maybe open an issue and include a code snippet. I figured a pull request was the "right" way to do it, but I know that you aren't exactly asking for contributions, so it's fair if you just want to ignore this. I really like the tools that this project helped me learn about, though, so I wanted to give something back. In fact, I have another little change in mind that can cut that ~5s rebundle down to <1s if you're open to it.

#### Notes:

[1]: The `vendorJS` task includes a pipe to `uglify()`. Uglifying 60,000 lines of libraries makes the original `gulp build` take longer. Commenting that makes the process take 14s.
[2]: Link: https://github.com/substack/node-browserify#bexternalfile
[3]: Link: http://benclinkinbeard.com/posts/external-bundles-for-faster-browserify-builds/
[4]: Link: https://github.com/sogko/gulp-recipes/blob/master/browserify-separating-app-and-vendor-bundles/gulpfile.js
[5]: Link: http://9elements.com/io/index.php/external-bundles-with-browserify-and-gulp/