@goodbomb 2015-04-07
I'm totally open to contributions. I just want to understand exactly what is being contributed before I merge a pull request. A PR is the best approach for sharing your approach as well IMO, so good call on that. This way we can have a discussion about the code that you're contributing instead of just an idea.

After your explanation, I'm leaning in favour of your argument. I also had the chance to sleep on it and think about it some more and I'm liking the way this sounds. This approach also helps keep the size of the development code base down to a minimum, and since that's what will be changing most frequently, it benefits the end user as well.

> I really like the tools that this project helped me learn about, though, so I wanted to give something back. In fact, I have another little change in mind that can cut that ~5s rebundle down to <1s if you're open to it.

I'm glad you like the project and got something out of it. That's what open source is all about :). I'm definitely open to further revisions if you have an idea that will improve things even more. Why don't you commit your changes and we can take a look at the PR once you're satisfied with your code, at which point we can make further revisions if necessary and then merge the PR.