@leftiness 2015-04-07
As you can see in that commit <sup>[[828505e](https://github.com/leftiness/angular-gulp-browserify-starter/commit/828505e7c3af6a3ec7c451c35a9703372e15486e)]</sup>, all I did is remove the gulp.watch() on the JS files. When there's a gulp.watch() on them, gulp runs the entire bundle-dev task whenever there's a change. Even with the external bundle that I added earlier, that still takes about 5s. If you instead let Watchify handle that (lines 160/186 of the gulpfile), then it takes <1s to rebundle the JS. That's really all there is to that. Watchify is really great.