@goodbomb 2015-04-08
Just tried it out locally. I like it a lot. Great job. You've got a typo in the lodash path, but it's no big deal. I'll merge it and fix it myself locally first and then push the changes.

Again, great work. Thanks a lot for the contribution!